+++ py_src/snex/__init__.py
@@ -1,5 +1,15 @@
 from .etf import set_custom_encoder
 from .interface import call, cast, send
+from .logger import LoggingHandler
 from .models import Atom, DistinctAtom, Term
 
-__all__ = ["Atom", "DistinctAtom", "Term", "call", "cast", "send", "set_custom_encoder"]
+__all__ = [
+    "Atom",
+    "DistinctAtom",
+    "LoggingHandler",
+    "Term",
+    "call",
+    "cast",
+    "send",
+    "set_custom_encoder",
+]